refactor(ScreeningDetailActivity): 뷰에 데이터를 설정하는 로직을 액티비티가 아니라 뷰 홀더에서 수행하게 함으로써 액티비티의 코드를 줄임
